# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Event-driven notification system for SMS, Email, and Push channels",
        "title": "Notification System API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/notifications": {
            "get": {
                "description": "List notifications with optional filters and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notifications",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by channel",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by start date (RFC3339)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by end date (RFC3339)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new notification request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Correlation ID",
                        "name": "X-Correlation-ID",
                        "in": "header"
                    },
                    {
                        "description": "Notification request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Notification"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/batch": {
            "post": {
                "description": "Create up to 1000 notifications in a single batch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create batch notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Correlation ID",
                        "name": "X-Correlation-ID",
                        "in": "header"
                    },
                    {
                        "description": "Batch notification request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BatchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.BatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/batch/{batchId}": {
            "get": {
                "description": "Retrieve all notifications in a batch by batch ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get batch notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batch ID",
                        "name": "batchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Notification"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/{id}": {
            "get": {
                "description": "Retrieve a notification's details and status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notification by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Notification"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/{id}/cancel": {
            "patch": {
                "description": "Cancel a pending or queued notification",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Cancel notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/templates": {
            "get": {
                "description": "List all notification templates",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "List templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Template"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new notification template with variable substitution support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Create template",
                "parameters": [
                    {
                        "description": "Template",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Template"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Template"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/templates/{id}": {
            "get": {
                "description": "Retrieve a template's details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Get template by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Template"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.BatchRequest": {
            "type": "object",
            "required": [
                "notifications"
            ],
            "properties": {
                "notifications": {
                    "type": "array",
                    "maxItems": 1000,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/model.CreateRequest"
                    }
                }
            }
        },
        "model.BatchResponse": {
            "type": "object",
            "properties": {
                "batch_id": {
                    "type": "string"
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Notification"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.Channel": {
            "type": "string",
            "enum": [
                "sms",
                "email",
                "push"
            ],
            "x-enum-varnames": [
                "ChannelSMS",
                "ChannelEmail",
                "ChannelPush"
            ]
        },
        "model.CreateRequest": {
            "type": "object",
            "required": [
                "channel",
                "content",
                "priority",
                "recipient"
            ],
            "properties": {
                "channel": {
                    "$ref": "#/definitions/model.Channel"
                },
                "content": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/model.Priority"
                },
                "recipient": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.Notification": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "integer"
                },
                "batch_id": {
                    "type": "string"
                },
                "channel": {
                    "$ref": "#/definitions/model.Channel"
                },
                "content": {
                    "type": "string"
                },
                "correlation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "last_attempt_at": {
                    "type": "string"
                },
                "max_retries": {
                    "type": "integer"
                },
                "priority": {
                    "$ref": "#/definitions/model.Priority"
                },
                "provider_message_id": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.Status"
                },
                "subject": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "variables": {
                    "type": "string"
                }
            }
        },
        "model.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.Priority": {
            "type": "string",
            "enum": [
                "high",
                "normal",
                "low"
            ],
            "x-enum-varnames": [
                "PriorityHigh",
                "PriorityNormal",
                "PriorityLow"
            ]
        },
        "model.Status": {
            "type": "string",
            "enum": [
                "pending",
                "queued",
                "processing",
                "sent",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusQueued",
                "StatusProcessing",
                "StatusSent",
                "StatusFailed",
                "StatusCancelled"
            ]
        },
        "model.Template": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "channel": {
                    "$ref": "#/definitions/model.Channel"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "variables": {
                    "description": "JSON array of variable names",
                    "type": "string"
                }
            }
        }
    }
}